+++ src/main/java/com/sdl/selenium/web/utils/FileUtils.java
@@ -128,7 +128,7 @@ public class FileUtils {
 
     public static boolean deleteFile(String filePath) {
         File file = new File(filePath);
-        return (file.delete());
+        return file.delete();
 
     }