+++ release-notes.md
@@ -1,6 +1,7 @@
 RELEASE NOTES
 -------------
 **Release Notes for Testy 2.1.0-SNAPSHOT**
+- add support for css selector when setExcludeClasses
 
 **Release Notes for Testy 2.0.0**
 - Change config default value for weblocator.log.suggestions=false